Added missing backticks (#9)
doccommentstyle.md was missing backticks on the 6th line, which made ` /* ... */ vs /** ... **/ ` appear as ` / ... / vs / ... / ` because of markdown's syntax regarding asterisks. To fix this, there should either be an escape character before each asterisk or the example should be in between backticks. Since other instances of asterisks were in between backticks in the rest of the documentation, I have added backticks instead of escape characters for consistency.